+++ CHANGELOG.md
@@ -4,6 +4,10 @@ See what's new and what has changed in bkper-js
 
 ## 2026
 
+**August 2026**
+
+- Added `Book.getTransactionsByIds` to retrieve complete transactions by ID in sequential batches of up to 200
+
 **July 2026**
 
 - Added `Bkper.requestBookAccess`